Fix FFmpeg libx264 AVC-Intra encoding in mov container
FFmpeg when encoding AVC-Intra in mov with libx264 sets b_repeat_headers to 0
and calls x264_encoder_headers() which calls x264_sei_version_write(). Neither
is allowed in AVC-Intra. Handle this by ensuring that b_repeat_headers is
always set and that the version SEI isn't written when in AVC-Intra mode.

Signed-off-by: Thomas Mundt <loudmax@yahoo.de>